tests/wpt/tests/css/css-view-transitions/transition-skipped-from-invalid-callback.html

<!DOCTYPE html>
<html>
<title>View transitions: transition skipped in animation phase</title>
<link rel="help" href="https://www.w3.org/TR/css-view-transitions-1/">
<link rel="author" href="mailto:khushalsagar@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  width: 100px;
  height: 100px;
  background: blue;
  contain: paint;
  view-transition-name: target;
}

::view-transition-group(target) {
  animation-duration: 300s;
}
</style>

<div id=first></div>

<script>
promise_test(async t => {
  assert_implements(document.startViewTransition, "Missing document.startViewTransition");
  return new Promise(async (resolve, reject) => {
    let transition = document.startViewTransition(() => {
      let foo = bar;
    });

    let readyRejected = false;
    transition.ready.then(reject, () => { readyRejected = true; });

    let domUpdateRejected = false;
    transition.updateCallbackDone.then(reject, () => { domUpdateRejected = true; });
    transition.finished.then(reject, () => {
      assert_true(readyRejected, "ready not rejected");
      assert_true(domUpdateRejected, "dom update not rejected");
      resolve();
    });
  });
}, "transition skipped because callback has invalid syntax");
</script>
