LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/transition-skipped-from-invalid-callback-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: bar

Harness Error (FAIL), message = ReferenceError: Can't find variable: bar

PASS transition skipped because callback has invalid syntax
